pyproject.toml
[tool.poetry]
name = "new_year_congrats"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
loguru = "^0.5.3"
python-telegram-bot = "^13.9"
python-dotenv = "^0.19.2"
colorama = "^0.4.4"
SQLAlchemy = "^1.4.29"
psycopg2 = "^2.9.2"
pygsheets = "^2.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
